requirements.txt
python-telegram-bot==20.8
openai>=1.50.0
aiosqlite>=0.20.0
python-dotenv==1.0.1
colorlog>=6.8.0
